+++ src/yt_config_output.c
@@ -300,7 +300,7 @@ yt_config_genesis_valid(float threshold)
 }
 
 bool
-yt_config_compose_hq_prompt(float current_hq, float upper_bound,
+yt_config_compose_hq_prompt(float current_hq, uint16_t upper_bound,
     size_t initial_column, struct yt_config_output_result *result)
 {
 	static const uint8_t newline = '\r';
@@ -318,7 +318,7 @@ yt_config_compose_hq_prompt(float current_hq, float upper_bound,
 		return false;
 	if (!append_literal(result, "Location? [ 8 to "))
 		return false;
-	if (!append_single(result, upper_bound, true))
+	if (!append_single(result, (float)upper_bound, true))
 		return false;
 	return append_literal(result, "] -=> ");
 }
@@ -343,9 +343,9 @@ yt_config_compose_hq_diagnostic(enum yt_config_hq_diagnostic diagnostic,
 }
 
 bool
-yt_config_hq_in_range(float candidate, float upper_bound)
+yt_config_hq_in_range(float candidate, uint16_t upper_bound)
 {
-	return candidate >= 8.0f && candidate <= upper_bound;
+	return candidate >= 8.0f && candidate <= (float)upper_bound;
 }
 
 bool
